renamed VanillaGameRuleSynchronization to GameRuleSynchronization and extended javadocs to better explain what to use the class for added license headers
